Fix filters with quotes in them
When we fixed multiple filters using quotes it broke filters with quotes in
them and I didn't notice because I was bad about running integration tests.
This fixes both.

I'm really not clear on why \" and \\\" don't work and the lookbehind does
work but I'm tired of trying to puzzle out multi-escaped escaping escape
characters.

Change-Id: I9ad1854adee266fd2b1d827c7bc18a51f2f2cdf0