Prefix search always squashes accents
Enable accent squashing for all prefix searches.  The argument is that
even if you speak a language that prefers accents not be squashed you
might be on a keyboard without accents.  We're not taking this argument
further and adding accent squashing to all searches at this point (or
maybe ever).

Bug: 67521
Change-Id: I1e9dc357b861edf1ff3c959e3ac3b287292c2bea