Allow limiting morelike to pages with wikidata
This adds a new prefix, "morelikewithwikibase:", which work just like
"morelike:" but filters out results without wikibase items.  For this
to work we index the wikibase item id if one exists.  This is safe even
if wikibase isn't installed because, well, we'll never put anything in the
field.

NOTE: I didn't add an integraiton test for this as I don't want our
integration tests depending on wikidata at this point.  Maybe we'll take
that plunge later, but not today.

DEPLOYMENT: This will always fail to find anything until the index has been
rebuild fully - both with a schema update and a full, from mediawiki, reindex
to pick up with wikibase_item.

Change-Id: Ic7eec0027a1a6216a02c87c5810a99dd3dd9b3e9