Create offsets for redirects
This should speed up highlighting for pages with many many many redirects
at the cost of a little disk space.

DEPLOYMENT: This requires an in place reindex to take effect.
DEPLOYMENT: Probably worth doing on wikis that we've seen slow prefix
searches.  This is certainly the cause of some of that and might be the
cause of all of it.  Wikis are eswiki and frwiki.  May as well do dewiki
and enwiki too while we're at it.

Change-Id: I1eacfcb41b622b9ed0462acb69cb4a6b96cf8912