Per user request limits
Needs poolcounter's Ie282b8486c7bad451fbc5fb9a8274c6e01a728a7 before we can
set $wgPoolCounterConf['CirrusSearch-PerUser'] otherwise we just abuse the
poor poolcounter.  If we don't set $wgPoolCounterConf['CirrusSearch-PerUser']
then this is just a noop.

Adds some testing code that detects if the poolcounter is setup and configures
cirrus to use it if it is.

Change-Id: Iedda2d1f39bc70bb8103d0fdfb1297708c960766