Make Jenkins helper work with MediaWiki-Vagrant
The Redis server used on a MediaWiki-Vagrant install does not use
password auth. RedisConnectionPool attempts to authenticate when
connecting with Redis unless the configured password is null. Since this
file is forced to load when the CirrusSearch role is enabled in
MediaWiki-Vagrant it seems like the configuration it forces should work
there.

Change-Id: Ie0f8326f0881aedc2ea7fff5167f7e532fdecde7