Support ORs of incategory
This supports incategory as names of the category or as the page id of the
category's wikipage.  The latter is to support integration with CatGraph.

This patch has a non-strict dependency (nothing will break without it) on
I7b7be24f in core. Until that patch is merged users searching for a
non-existent category by id will be offered to create the page matching the
search query when they should not.

Bug: T89823
Change-Id: Iad60c33d6a582a1ad005bd45165ffa198637add0