Add support for phrase_prefix match queries
This alters search terms that match the pattern /^"([^"*]+)([*])/
to be phrase_prefix match queries rather than components of a
query_string query.

Specifically, this is intended to fix the following case:

  "Феофан Затворни*" >> 0 results

from https://phabricator.wikimedia.org/T88724

Though this change only covers a single case of the above pattern,
it sets up a foundation that can be extended to cover all phrase match
queries (including support for slop, negation, etc.) in subsequent
commits.

Change-Id: I0da005fa5d994642cc53c6cfb482efe8b977813a