Improve highlighting for phrase_prefix queries
This fixes addresses https://phabricator.wikimedia.org/T93014 to
enable highlighting in the title, etc., as well as for compound
searches including things other than phrase_prefix queries.

This relies on a sorta hacky query_string query for the highlighter,
because when we use a match query only the title gets highlighted.

Change-Id: I80cbd3c7daacad8680f2a32ab91fac12d915a7ed