Don't trigger Wikitext highlighting for intitle:
This addresses https://phabricator.wikimedia.org/T94870

In f587e00919ad1cbb5be43b2f54ad7777ac02b739, we lost a conditional
that checked whether an {insource|intitle} term was insource, and if
so use Wikitext for highlighting.  This restores the lost contional,
and prevents triggering this condition for intitle terms.

Change-Id: I722e25d6c5c07da5f23080fe278d95cb59c2e018