Move Gemfile from root to browser tests
The gemfile refers to the dependencies necessary for the browser
tests to run, and have no direct bearing on the main project.
Additionally the puppet code for vagrant browsertests expects to
find and install ruby dependencies from the tests/browser directory.
This change is required for vagrant puppet to actually install the
dependencies so that tests can run inside vagrant or on a vagrant based
labs instance.

Change-Id: I3795b9f921b17d3509dbf08baee32ea4717b143b