Revert "Move Gemfile from root to browser tests"
This reverts commit 4ec9068d4f5a4294ad2cf0fd437abebd12b8b3b6. There
are still issues to be resolved in the CI infrastructure for this
to be properly merged.  I still think it is correct to move this
into tests/browser for the reasons mentioned in the original patch and
will revisit when i have a chance to review how the CI infrastructure
works in this area.

Change-Id: I98d47d4d4cefafd9a13502a6322f329f2ba2e317