Rework regex for quoted keyword:"..." extraction
I'm touching this code for two reasons:

1. A trim() to remove the quotes was found in three places. We can
remove the quotes right away in a single place. I think this makes
the code easier to read and maintain.

2. A simple trim() will lead to unexpected results in cases like
keyword:"\"foo\"". The escaped quotes should be part of the search
string but trim() will remove them because trim() can not be limited
to trim only 1 character left and right.

Change-Id: Id6f58010c5a69bd20090bc2c400cf2b93c8da543