Fix intermittent failures in prefer_recent tests
The final prefer recent test of .4,.0001 requires about 13 seconds
after editing, locally, for PreferRecent Second Second to win out
over PreferRecent Third.  This isn't the time to index, but the time
necessary for the weighting to meet our expectations.  Verified by
repeatedly searching in a browser after performing the edit.

Change-Id: I87579097a9e6cc485dff0998154480bea122027b