Avoid counting arrays if not needed
count() tends to be one of the slowest PHP functions and should be
avoided, especially if the actual number of elements is not needed.
empty() is well defined for variables that are guaranteed to be an
array (or null).

empty() should be avoided on strings, but this is not the case in
this patch.

Change-Id: I5e01fe5d3fbcf0f32e602283075bb5967748c02a