New option to log but not fail per-user pool counter failures
When turning on the per-user pool counter in production we had a significant
number of hits almost immediately. Patch adds a new feature flag that will
log but not fail the queries when they hit. This should help us to tune the
settings and figure out the right values before we start blocking queries.

Bug: T76497
Change-Id: I7e1b8c3411cbbcc9d39b6e9031db9c8975e26779