Log the username for pool counter failures
To figure out if ip addresses need to be whitelisted or are being appropriatly
blocked we need to know what the ip address is. This adds them to the log. While
the IP is PII mediawiki is already logging that ip address all over the place.
We have no need for non-ip usernames, so those are not included in the log.

Change-Id: I698353a94cd4c4bc8b404483605b25d315b16657