Support pausing writes to elasticsearch
Creates a cluster wide index, mediawiki_cirrussearch_frozen_indexes, which
tracks which indexes are currently frozen and do not allow writes. Supports
freezing both individual indexes and all indexes cluster-wide.

To support the pause operation all writes are shifted into Job classes. When
the write operation is first desired the job is created and then run in process.
If the job detects that the desired indexes are currently frozen it will inject
itself into the job queue. Future runs of the job will continue re-injecting
themselves into the job queue with an exponential backoff until writes are
re-allowed.

Bug: T99244
Change-Id: Ief7019f314c74e8a167505663bdfcec6c43c4567