Suppress interwiki results when they would break
Sometimes we end up with a SearchResults object that can't handle interwiki
results and it just isn't fair to try to feed them to it. It makes fatals.

There is another fix for this issue, but its more complex and as of yet
doesn't pass all the tests. But this is a simpler, more obvious fix we can
can use as a stopgap.

Bug: T104189
Change-Id: Id46545dd77a00ccd4ba2f18efc5d7d57f2d96626