Include session_id in CirrusSearchRequest logs
We would like to tie together multiple requests made by the same
browser within a short timespan. This unfortunatly doesn't have
a time limit to matching, but generates a non-identifiable token
that is stable per-user by hashing the ip address, xff header and
user agent.

This will allow us to calculate stats such as number of searches per
user. It will also allow us to see if users that get zero results for
a prefix search continue on to get positive results for the same or
similar query in full text search (happens automagically on mobile, and
by pressing enter from top right search in desktop).

Bug: T106552
Change-Id: I3493947fbc96689c5cf820475f3a701549454273