Move query rewriting into search backend
This is the CirrusSearch half of I0a8f75759. In that patch rewriting
was adjusted so it can be implemented directly by the search engine.

This will allow us to make more indepth changes to the way the second
tquery is run, rather than only having control over the text of the
search.

Bug: T106888
Change-Id: Iebce6a59931ae0c295e13ccc12a22a1ae2081777