Bare bones helper for AB testing
This is a simple bare bones class for determining if the current
request is a member of an A/B test. It provides the same experience
across multiple requests by calculating the bucket from a hash of
the test name, ip address and user-agent. It doesn't do anything
particularly fancy yet, that can be built out later.

Did some TDD to write this...but i think it turned into too many tests.
Can prune later i suppose.

Bug: T106888
Change-Id: Idd55b3e065fbb5e8248f406580f935545678ae37