Normalize CirrusSearch in stats collection names
It might be nice to be able to wrap a stats collector
to prefix everything, so in one place we wrap to
core collector, and everywhere else is guaranteed
to use the same prefix. But there are only two for
now, so thats probably overkill.

This just changes one string.

Change-Id: Iac60d86908416fd23cec511dd5282b7d41d3fe86