Run alt language search if suggestions has no results
Currently we run the suggestion if it was available, but
when that returns no results we just give the user the no
results. Lets try out the language search in the case.

This was prompted by my tests to get german queries to
show german results in english wikipedia, enwiki has enough
german words sprinkled about that the suggester gives a
suggestion, but in numerous tests (by hand, random queries
from our logs) i was getting no results on those suggestions.

Change-Id: I8f7d44f330f99c16aef045da8db0f9ce90c0f937