Don't run maintenance scripts for Elasticsearch 2.x
ES 2.0.0 is a major release, which may have breaking changes, that
aren't supported in CirrusSearch/Elastica (currently). One thing is,
e.g., that the _status api was replaced by Indicies recvery[2] and
Indices stats (_stats endpoint)[3], see [1].

This change results in an exception in Elastica, which, in the latest
release, supports ES 1.7.3, only. The maintenance sccript should exit
as early as possible, instead of trying to use ES 2.0.0 and fail with
a user-unfriendly exception.

[1] https://www.elastic.co/guide/en/elasticsearch/reference/2.0/indices-status.html
[2] https://www.elastic.co/guide/en/elasticsearch/reference/2.0/indices-recovery.html
[3] https://www.elastic.co/guide/en/elasticsearch/reference/2.0/indices-stats.html

Change-Id: I52e93ea2c1b5f602dc101c8c79fd958fbb8a941c