Add option to urldecode queries in runSearch.php
While running queries extracted from web request logs I ran across
the issue that some queries contain \n in them. We could pre-normalize,
but that's not what production does. In a quick review it looks like
SpecialSearch strips the \n, but ApiQuerySearch does not. A followup
patch strips the \n in cirrussearch so we give the same results
in both places.

Change-Id: I63f7c7337f1dc09d4c32d00fba8e5ada1f4a2701