Disable check for Jenkins.php browser tests entry point
afaik, nothing uses the entry point now.

Cindy the browser test bot always directly includes the
Jenkins.php regardless of this check in the entry point.

Jenkins has some browser tests that run on beta wikipedia:
https://integration.wikimedia.org/ci/view/BrowserTests/view/CirrusSearch/

But the php code is run on beta and not on jenkins, and
thus the entry point is not needed for those.

Without removing or making the check more specific,
other jenkins jobs (e.g. Wikibase) that do not have all
the optional extension dependencies (e.g. MwEmbedSupport)
enabled and then those jobs fail.

This makes the check more specific to possible CirrusSearch
browser test jobs that might run on Jenkins and need it,
and disables the check for now. It would be nicer to
not need this special entry point at all, if possible. :)

Bug: T120360
Change-Id: Ieab7d72f005db3d1304014ef7620e44479564c02