Add more flexible rescore profiles
This patch adds more flexibility to the rescore profiles. In the previous patch
it was impossible to adjust weights for multiple function score.
Added a new function type 'script' to use lucene expression script engine.

Added the possibility to specify a different profile for prefix search or
morelike queries. Useful to use a complex profile for fulltext search that may
not be suited for fast prefix search or morelike where we do not control field
boosts.

Bug: T110648

Change-Id: I8b3639317dd86a9cde9633ef490e48dfe4fa4d4b