Include language detection results in control bucket
This is necessary to properly report the control buckets to
the user testing logs. Without it analysis would have to
manually reconstruct which control users could have been
in the test which, while possible, is not going to be as
provably accurate.

For users not participating in the test
$wgCirrusSearchLanguageDetectors will be an empty array
so this should result in unnoticable latency increases.

Change-Id: Ic885be900209eca7b15f2f2fa47036e1de082596