Utilize the master timeout for creation and deletion as well
Unsurprisingly, mapping updates are not the only things that
can timeout. This applies the master timeout to creation
and deletion of the indices as well.

In addition allow specifying the timeout on the command line,
this can help figuring out what the right timeout is.

I thought about applying all this to the standard update
process as well, but it's simply too many hacks. I will
work out a patch to make the rest of these available in
Elastica library, and we will update after it is released.

Bug: T107348
Change-Id: I8acf109f1bcab83356d7164b46c2ef1c522e1ccd