For source code pages, index page contents as `source_text`
Pages containing source code or machine-readable data (like
MediaWiki:Common.js) currently have no 'source_text' value.
The source code is emitted as the 'text' value. As a result,
'insource:' queries never match JS / CSS pages, which is both
counterintuitive and annoying.

So do as Nik suggested in T88247 and make the 'source_text' of
source code pages be the same as the page text.

This causes the same body of text to be emitted twice in the page
dump. But since these pages represent a very small fraction of
Wikimedia content, the storage costs are neglible.

After this change rolls out, something will need to trigger a
re-index of JS / CSS / JSON pages. It can either be a normal
edit to the page, or we can force it with a script.

Bug: T88247
Change-Id: Ibae6e02cd597de5d4f0efbfff8321aabf262d426