Don't sanitize non-wikitext going into search
As far as I can tell this sanitization was added back when
CirrusSearch was being written for solr cloud. I've tested
a few different methods and afaict the experimental highlighter
along with the normal elasticsearch highlighter both properly
encode the highlighted output and this is unnecessary.

Currently only removing the Sanitizer call from the `text`
field. After reviewing this I don't believe there to be any
security concerns around removing the Sanitizer, but it could
have some effect on search results and highlighting output that
would start to have html tags that might only confuse the user.

This had the unintended side effect of stripping non-tag content
from javascript, causing some searches for content of non-wikitext
pages to be incorrect.

A full reindex of all non-wikitext content will be required
after merging this patch.

Change-Id: I0706ce3f178ad791bdd537fa88d6454f6ac9ae15