Allow redirecting more like this to a different cluster
Eqiad is showing latency spikes on the weekends that we suspect
are due to increased load from increased usage of the more like this
feature.

Using this will add some round trip latency to more like this queries,
but we suspect will even out search latency across the board under
heavy load situations.

One thing not addressed here is these still use the eqiad pool counter,
as opposed to one specific to codfw. That's likely an acceptable
problem, but one that should be addressed eventually in a more 
complete manner than a single query.

Bug: T124100
Change-Id: I0051461f1ca2bbd1c74e3b6a6f86e5e289a8fe27