Cache more like queries into ObjectCache
I'm not incredibly thrilled about how this works, but I think it will
get the job done. There are so many options and ways for queries to
vary I had to take a brute force approach of serializing things
that look relevant to the result and md5'ing a combined string.

Potential issues with this solution is that we don't try particularly
hard to normalize queries. About the only thing that was done is
re-ordering a few arguments that have no effect on the query result
afaik.

This additionally adds a few new statsd variables under
CirrusSearch.query_cache.more_like so we can monitor how this is working
out.

Bug: T124216
Change-Id: If475db863465fc51b747b5760020453ef3417ea5