Access first heading by class name, not by id
It does have an id and a class name when running the tests with the
Vector skin enabled, but does not have an id when the skin is disabled.
We (the Wikidata team) run into problems today because of this. I
suggest to do the same changes we did to our tests here.

Change-Id: Ic08112fab9de66d8418331d8870609c0f2734374