Try to cleanup broken indices when building comp suggest indices
In some case a broken index can survive a failed update, this patch tries to
clean up such indices instead of failing asking for manual intervention.  If
the broken index is not referenced by an alias and not suggest are reported by
the stat API then it's safe to remove and continue.

Bug: T127182
Change-Id: I677b265cfcf6f4a4bc28994d29a403c61f660b77