Remove cirrus-suggest api and supporting js
This is deprecated in favor of recent work to trigger completion
suggester directly from the main opensearch api. Anything that
could have been done with this is now doable from the standard
mediawiki core actions.

This was an internal API call and only used from the js file here,
it should be safe to completely remove.

Change-Id: I051437a5ee238716dd773ee66951ce25bef984b1