Add unique id and page ids returned to CirrusSearchRequestSet
This is something we had talked about for building click through logs to
use in rel forge, or other forms of improving search based on prior user
behavior. It adds a unique id to the request set log, and also adds that
unique id to the search url. This allows extracting search click
throughs from webrequest logs and associating them with a particular
search. By including the returned pages in the request set log we can
also determine which position the user clicked through on. Along with
some sessionization in hive this should be enough to build up data sets
for scoring purposes in rel forge.

As long as the CirrusSearchRequestSet definition is being updated this
also moves some items from payload into the main schema.

Change-Id: I0ddd4da00c35ca5883235871674482a9a570bd8b