Added Similarity configuration
* User can now configure the similarity function and parameters used by the
  index
* Re-enabled norms on arrays fields

Norms were explicity disabled on array fields which maybe makes sense with the
default similiraty, docs with a high number of redirects were maybe ranked too low
but without norms it's the opposite, used with a DisMax the redirect field will
always win thus docs with a high number of redirects are likely to win.
I think it's better to re-enable them and allow the user to fine tune a parameter
that affects norms (setting a very low b value with bm25).

Bug: T128061
Change-Id: Ie718c8a03ea5351e2e142f22436e2fd83314056f