CompletionSuggester: add support for crossnamespace redirects
The suggester index is now built from content and general index.  It filters
docs with a redirect from the main namespace.
Removed ':' from the list of ignored char, ignoring ':' can be misleading for
certain kind of usage:
* wikipedia shortcuts like WP:WP when WP is not an alias to Wikipedia

Note that recycling won't be possible for the next update because the analysis
config changed.

I'm not a big fan of the way I deal with LinkBatch...

Bug: T129575
Change-Id: I77779fe066409c62ee5c8b3bf936500e6924b23a