Add temporary code to investigate why we receive unknown titles from SearchEngine
This is bugging me, I can't reproduce the issue locally and apparently it
happens on more than 300qps (according to logstash).
We should understand what's going on here because 300qps can have a significant
impact on all analysis using CirrusSearchResultSet.
Hopefully with the help of the new queryString payload we will be able to
identify the cause?

Change-Id: I873b5e7bf76faefcf2e6e3f42de9b75e6e68fac5