Standardize on catching/passing Elasticas ExceptionInterface
We catch elastica exceptions in various places and pass them around, but
were not consistent on if we typed on Exception or ExceptionInterface.
This standardizes to ExceptionInterface.

There is one downside of this patch ... Elastica's ExceptionInterface
does not implement any methods. This means we can't type check any
method calls against the objects...Sadly the newest versions of elastica
don't rectify this situation.

Bug: T132625
Change-Id: I0399bb8919d08c5b5fb7a15e3e62a35156ff74f3