Dont raise warnings on misconfigured morelike settings message
If the message is malformed and doesn't strictly adhere to "foo:bar"
formatting we can end up with quite a few warning notices in the logs.
Instead of raising all those warning's just gracefully handle a
malformed string.

It would be nice if we had some sort of way to validate the message on
save, both at translatewiki and if local admins change the message, but
I'm not sure how to accomplish that.

Change-Id: I1d22a13bae4fe32b7c08267d8c538a6c2d167c07