Remove phrase suggester tests based on term frequencies
max_term_freq and min_doc_freq are not working as expected.
These tests used edge values like 0.0000001 and 0.9999999
but stepping threw eclipse debugger I noticed that these values
were read as 0 and 1. Testing on elastic 1.7 and enwiki these settings
have absolutely no effect.
Reading some code in elastic it seems that candidates with term freq set are
only loaded for the first tokens emitted as a synonym:
https://github.com/elastic/elasticsearch/blob/2.x/core/src/main/java/org/elasticsearch/search/suggest/phrase/NoisyChannelSpellChecker.java#L86
So all other candidates are generated with a termFreq of 0 thus disabling
the effect of these settings.

Finally: the purpose of the test was to make sure that these values are passed
to elastic, unfortunately without any change in the response/behavior it's
impossible to test it.

So howto explain why these tests worked before: I don't really know, I suppose
it's because 0.000001 and 0.999999 were tranformed to 0 and 1 and we entered
into some corner cases in the code that depend on a particular index states.

Bug: T133747

Depends-On: Ida99848eb27516b60b343f6ca5bcf140ad46a472
Change-Id: I3e87934e30186447e3c42e101d9b5a2b842d02c3