Drop support for common terms query
This was an experiment that didn't end up making a useful difference in
search. The feature is unused so lets clean it up.

There were some other refactors made when adding the common terms query
arround how querys get build (SearchContext, SearchTextQueryBuilderFactory
and such), but they seem useful enough to leave in and not also undo.

Bug: T133749
Change-Id: I10acc647008dc28b2fc33279180b0f6eb462fb4c