Fail and log error if page was not properly indexed
When errors come from the nginx proxy, rather than from elasticsearch
itself, the Elastica library does not notice that there is a problem
because it depends on elasticsearch providing an appropriate error
rather than by checking the http response code. To do a better job
handling this the Elastica library needs to be updated, but this patch
will work for now to notice and log errors about the failure to index by
noticing that no bulk responses were returned from elasticsearch.

Bug: T132763
Change-Id: Ib218112bce9fbca09d913550a60f738b8d7140a2