Simplify run-phan.sh by using cloudflare/docker-phan
Our current solution just felt hacky and had problems with the php7
installations conflicting with the primary php5 installation. Lets just
skip that whole kerfuffle and use a docker container that has phan built
in instead. This container is ~20M and is pretty simple to use. It does
take a few minutes to build the first time around though.

Also adjusts the script to not output random extra data, and instead
output the error given by phan. The exit code is now 0 if everything is
happy, or 1 if there are problems. This makes the script easier to
integrate into cindy for auto-running static analysis.

Change-Id: I9384648d15c5692f3d85f85705bc88ea2aa80b37