Hack together a line-specific @suppress annotation for phan
It was rather annoying to not be able to supress specific lines with
phan, so added a hacked together post processing script that will do it
for us. This is less than ideal, but unfortunatly php strips these
internal comments out of the AST before providing it to Phan so there is
not great solution that can be applied to phan directly.

Also adds a number of annotations, most with explanations of why, to
suppress errors that are ok. Mostly these are due to weaknesses of the
phan analysis phase that will probably improve over time.

Change-Id: I1bb9ce73370ae960fb6f6b10db2a01ffe01b0b29