Convert to short array syntax
Core and other extensions have already converted. We've been avoiding
this because it mucks with patches in progress, but at some point we
have to pull the trigger. Is now better than some other time? dunno.

Converted with
https://github.com/thomasbachem/php-short-array-syntax-converter by
running:

  find includes/ maintenance/ profiles/ tests/unit scripts/ CirrusSearch.php \
      -print0 | xargs -0 -L1 php convert.php -w

Rebuilding this patch to deal with conflicts due to merges should be
fairly easy.

Change-Id: Ib74d78d4141e97c106589a2a6f0dc6b49b97cc64