Automatically select sanitization settings based on wiki size
The following settings are tuned based on a push refresh rate at 2 hours.
Avg loop duration is 4 days, 17 days max. It will increase joqueue throuput by
17 jobs/sec at the beginning and will decrease once some wikis will wait for
the min loop duration time set to 2 weeks.
It can generate an increase of 85 updates/sec per cluster if the number
of ids to fix is 18%.

wgCirrusSearchSanityCheck has been turned into a boolean that could
be used to activate sanitizations one wiki at a time for testing.

Bug: T139200
Change-Id: I1a2e6557c727f0ef7fad2de5be0778bcb9b14666