Refactor some state from Searcher::searchText to SearchContext
In preparation for refactoring parts of query parsing out of searchText
push some of the state into SearchContext. This should make it easier
to break parts out of the function.

This does start to explode the size of SearchContext, although it's
mostly boilerplate. Future work can further refactor things to resolve
that. One idea would be to move some of the methods that build pieces
of a query into one part into a QueryBuilder or something, and have
the context provide a getQueryBuilder() method instead.

Change-Id: Ie7a647163c4a9e0c6b61d461084cee31fe5d567e