Add a field to page documents identifying source wiki
To support storing multiple wiki's in the same index we need to identify
where a page actually came from. To allow this add a new field to page
documents indicating it's existence.

This is safe to deploy as is. We use `dynamic = false` for elasticsearch
mappings to until the mapping is updated elasticsearch will safely
ignore the new field until the mapping is updated. Once the mapping is
updated a full reindex will be required, but we can probably wait until
there is another reason before utilizing this new field.

Bug: T139500
Change-Id: I9d349f255614720a7b171f835e999be25da45ca5