Refactor forceSearchIndex.php to use iterators
This script was a hard one to read because there is lots of state that
gets shuffled arround to ensure the next set of pages starts at the
right point, and there were multiple methods for doing that iteration.

Refactor all that to use Iterator's to source the batches from. The
use of BatchRowIterator gets most of the state being maintained out of
this class and into a shared class that handles explicitly that.

Change-Id: Icc27b1e9a391535c685a0869ca48c52bd530be06