Centralize document id generation
To support multi-project indices we have to make sure document id's
don't overlap between projects. The main idea for this is to prefix
all ids with the wiki name, so the document 12 becomes testwiki:12.
Internal to elasticsearch id's are already treated as strings so
nothing needs to change there.

This patch tries to track down all the places we deal with id's and
classify them as document ids or page ids.  All the places page id's
need to be converted to pageId's they pass through SearchConfig::makeId().
This works, but I worry that it is fragile.

Enabling this and reindexing must be performed before doing any
kind of multi-wiki merging. By having the indices already configured
to use distinct id's the transition phase of reindexing into a
single index and cutting over should be worry free, requiring only
a standard time based reindex to cover updates during the initial
reindexing period.

Bug: T139496
Change-Id: I5f1aa56645f7ff88649e4238ca6d0af551c858c3