Speed up saneitizeJobs.php
On the production cluster fetching the status seems to take
extaordinarily long. Overall running saneitizeJobs.php --show takes 4 to
5 seconds. Similarly the cron job to insert saneitize jobs takes nearly
an hour and a half to run. Poking around it seems almost the entirety of
that time is spent waiting for elasticsearch to respond to the status
requests.  With just shy of 900 wikis in production, doing a foreachwiki
against saneitizeJobs.php takes over 60 minutes due to this.

Speed things up by using the Index::exists() call, which performs a HEAD
request against the specific index, instead of requesting the full
server status. In a quick test on terbium this brings the runtime down
from 5s to around .6s. Shaving 4s off each execution should reduce the
time it takes to loop through the wikis in cron by ~60 minutes.

Change-Id: I3e0936074bc7679f56be737a397b00ad3be44b15