Take SearchConfig directly in SearchTextQueryBuilders
These query builders were taking in a SearchContext, but the only thing
they use is info from the config itself rather than the state built up
into SearchContext. To make this clear, remove the dependency on
SearchContext.

This is a minor rework to make the next patch, extracting
Searcher::searchText into a FullTextQueryBuilder class, a bit simpler
and more direct.

Change-Id: Ic8f0b41ae257ca7913acafdf141c704eae54e9c2