Move Searcher::searchText() into dedicated class
Even after factoring out the keyword features out of this function it's
still excessively long. It's not the easiest thing to refactor directly
though due to the varied state being used. To make it easier to think
about pull the function and the methods only called by this function out
of Searcher and into a dedicated class. Also pulled the functions that
were previously moved into QueryHelper into the new class if they were
only used here. The function itself is mostly unchanged, followups will
try and cleanup the function itself.

Change-Id: I44447bb8347228f7a46f9b2b8728f3a223495064