Correctly build interwiki section titles
Looks like this one got missed at some point or another. We need to
provide the interwiki prefix to makeTitle so the titles generated
appropriately point to the other wiki.

Switched to Title::makeTitle rather than makeTitleSafe to be consistent
with other interwiki title creation within the class. The extra parsing
is unecessary here as the data was source not from a user, but from the
mediawiki database before being pushed into elasticsearch. Has the side
effect of allowing tests to run and pass without populating interwiki
tables.

Bug: T141502
Change-Id: Ic3943523b599054768cab2c39ffb039715eeac0e