Show interwiki namespace correctly in tooltip and title snippet
To show the interwiki namespace we have to fetch it from elasticsearch.
Add a basic check so if we are doing interwiki with FullTextResultType
we will also fetch that. The rest of the code finds this and works
as expected.

When creating interwiki titles the namespace id should not be used,
instead the namespace text should be prefixed to the title text. Handle
this appropriately as well. Unfortunately there are times (redirects
from other namespaces) that we don't have the namespace text, but these
should be relatively rare. This means there might still be some
inconsistencies with redirect titles, but this is better than it was.

Bug: T141503
Change-Id: I3f2710f8b758bad17a46b3a819781e4863ff8eb7