Add similarity config to suggest field and drop positions
This field is built outside the normal process but we still need to setup a
similarity config on it.
Refactored the similarity config function outside of TextIndexField so it can
be used everywhere.
Added index_options: freqs to save space by removing positions which are not
needed with this field.
Removed unused title.suggest and redirect.suggest subfields.

Change-Id: Ia115648aee7e134fccd37d0630d79499cd25637c