Implement a new fulltext query builder
As we plan to switch to BM25 the current query built on top of query
string and the all field is unlikely to work properly. The all field
deals with raw tf and due to aggressive term freq saturation done by
BM25 the problems identified with the all field and the classic
similiraty are likely to be worse.

The solution used by this patch is still a hack and uses
FullTextQueryStringQueryBuilder as a base to reuse its parsing logic.
If the query does not use any query_string features then it will
generate a set of simple match/multi_match elasticsearch queries.  It
fallbacks to QueryString otherwize. It means that queries that include
phrases, boolean logic, wildcards and fuzziness will be ignored by
this new builder.

The new builder will read a profile and generate a query accordingly
allowing to play with settings to tranform the generated query.

I had to update some tests:
 - Reordered phrase rescore Words Test Rescore because the original
   query was ambiguous and the new top result sounded good to me.
   (c.f. tests/browser/features/full_text_api.feature:26)
 - Added one word to Relevancytestphraseviatext so that norms are
   equivalent with other pages in the same test
 - Removed one word tests/browser/articles/Relevancytestviaopening.txt
   same reason (norms)
 - Changed in autocollapse to inside autocollapse, stop words are
   now no longer required (certainly a QueryString limitation due
   to crossfield query style being confused by a missing word making
   it mandatory on the plain fields)

I had to lower the impact of norms on redirect fields to comply with
tests/browser/features/full_text_api.feature:93.

The profiles (field weight and bm25 settings) are specifically tuned
to comply with the current tests, I don't think these settings are
suited for real world wikis.

The proper approach (imo) would be to decouple parsing and builder
logic, a parser should generate a kind of AST that will be rewritten
into an elasticsearch query.

Bug: T128073
Change-Id: I26185a4efd9c777519392fe56d7707d2eb720452